+++ src/test/groovy/ru/vyarus/dropwizard/guice/cases/multicases/support/Feature.groovy
@@ -1,0 +1,11 @@
+package ru.vyarus.dropwizard.guice.cases.multicases.support
+
+import ru.vyarus.dropwizard.guice.module.installer.feature.eager.EagerSingleton
+
+/**
+ * @author Vyacheslav Rusakov
+ * @since 01.08.2016
+ */
+@EagerSingleton
+class Feature {
+}